attach route to sse
