+++ lib/app/pages/pay/binding.dart
@@ -1,6 +1,7 @@
 import 'package:dimipay/app/data/modules/payment_method/controller.dart';
 import 'package:dimipay/app/data/modules/payment_method/repository.dart';
 import 'package:dimipay/app/data/provider/api.dart';
+import 'package:dimipay/app/data/services/sse/service.dart';
 import 'package:dimipay/app/pages/pay/controller.dart';
 import 'package:get/get.dart';
 
@@ -9,5 +10,6 @@ class PayPageBinding implements Bindings {
   void dependencies() {
     Get.lazyPut<PayPageController>(() => PayPageController());
     Get.lazyPut<PaymentMethodController>(() => PaymentMethodController(PaymentMethodRepository(ApiProvider())));
+    Get.lazyPut<PayResultSSEController>(() => PayResultSSEController());
   }
 }
